evalengine/compiler: arrange in separate files
